Always take a ref for all event so that the object is only freed when the op is released.
Reset sal pointer held by all ops when destroying the sal object
Avoid attempting to send messages if the sal has already been destroyed

Improve documentation of conference participant list type core and
conference parameter attributes as it is only applicable to servers

Rework linphone_core_get_call_history in such a way that the call log
list held by the core is updated to match exacty what is stored in the
database and its members are reused as much as possible. Basically for
every call log stored in the database, the method will look if one is
already in the list and reuse it.

Fix test to account for call log server incoming messages and internal
event class changes

+++ tester/call_single_tester.c
@@ -2217,7 +2217,7 @@ static void cancelled_ringing_call(void) {
 
 	call_history = linphone_core_get_call_history(marie->lc);
 	if (BC_ASSERT_PTR_NOT_NULL(call_history)) {
-		BC_ASSERT_EQUAL((int)bctbx_list_size(call_history), 1, int, "%i");
+		BC_ASSERT_EQUAL(bctbx_list_size(call_history), 1, size_t, "%zu");
 		LinphoneCallLog *call_log = (LinphoneCallLog *)bctbx_list_get_data(call_history);
 		BC_ASSERT_EQUAL(linphone_call_log_get_status(call_log), LinphoneCallMissed, LinphoneCallStatus, "%i");
 		// Make sure the call log duration is 0 as the call wasn't connected
@@ -5816,7 +5816,7 @@ static void incoming_invite_with_invalid_sdp(void) {
 	BC_ASSERT_EQUAL(callee->stat.number_of_LinphoneCallIncomingReceived, 0, int, "%d");
 
 	logs = linphone_core_get_call_logs(callee->lc);
-	BC_ASSERT_EQUAL((int)bctbx_list_size(logs), 1, int, "%i");
+	BC_ASSERT_EQUAL(bctbx_list_size(logs), 1, size_t, "%zu");
 	if (logs) {
 		const LinphoneErrorInfo *ei;
 		cl = (LinphoneCallLog *)logs->data;
